refactor(frontend): rename Ethereum send destination component (#4103)
# Motivation

For consistency, we rename the component `SendDestination` to
`EthSendDestination`.

---------
